requirements.txt
s3fs
pandas
kafka-python
